pyproject.toml
[project]
name = "klasfotos-raam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.1.0",
    "pyqt6>=6.8.1",
    "qt-material>=2.14",
]
